Task/Location Allocation
Locations.csv is a list of longitudes and latitudes

The python file reads this, asks the user for input on threshold distance, number of tasks, etc.

It then assigns tasks according to the user input by randomly sampling from the dataset. The user defines how many tasks he wants under the threshold and how many he wants over it.

It also optionally prints and locally saves a distance matrix which calcualtes all the possible distances from the locations finally chosen. This is so that it is easy for the user to see the distribution of distances.